feat(dashboards): Render widget thresholds in Slack unfurl chartcuterie chart
The SLACK_DASHBOARDS_WIDGET renderer previously emitted only the data
plottables, so widgets configured with thresholds rendered without the
threshold mark areas / mark lines that visualizationWidget shows in the
dashboards UI. Append a Thresholds plottable when the widget has
max_values.max1 or max_values.max2 set, mirroring visualizationWidget.

Plumb the threshold plottable through a new extraPlottables param on
buildTimeseriesChartOption so SLACK_TIMESERIES stays untouched.
toSeries receives the same union options shape that
timeSeriesWidgetVisualization passes, so non-data plottables can pick
up theme / maxOffset while ignoring the per-series fields.

The Thresholds plottable imports normalizeUnit from
sentry/views/dashboards/utils, whose dependency graph drags
fieldRenderers.tsx (UI components, badges, dropdowns) into the
chartcuterie bundle. Move normalizeUnit to its own slim module so the
chartcuterie bundle doesn't inflate. Re-export from dashboards/utils
to keep existing FE callers working.

Fixes DAIN-1606
