+++ static/app/chartcuterie/dashboardsWidget.tsx
@@ -1,10 +1,13 @@
 import type {Theme} from '@emotion/react';
 
+import {defined} from 'sentry/utils';
 import {type Widget} from 'sentry/views/dashboards/types';
 import type {TimeSeries} from 'sentry/views/dashboards/widgets/common/types';
 import {formatTimeSeriesLabel} from 'sentry/views/dashboards/widgets/timeSeriesWidget/formatters/formatTimeSeriesLabel';
 import {formatTimeSeriesLabelForWidgetQuery} from 'sentry/views/dashboards/widgets/timeSeriesWidget/formatters/formatTimeSeriesLabelForWidgetQuery';
 import {createPlottableFromTimeSeriesAndWidget} from 'sentry/views/dashboards/widgets/timeSeriesWidget/plottables/createPlottableFromTimeSeries';
+import type {Plottable} from 'sentry/views/dashboards/widgets/timeSeriesWidget/plottables/plottable';
+import {Thresholds} from 'sentry/views/dashboards/widgets/timeSeriesWidget/plottables/thresholds';
 
 import {buildTimeseriesChartOption, CHART_SIZE} from './timeseries';
 import type {RenderDescriptor} from './types';
@@ -29,9 +32,26 @@ export const makeDashboardsWidgetCharts = (
       const queryIndexByTimeSeries = new Map<TimeSeries, number>(data.timeSeries);
       const flatTimeSeries = data.timeSeries.map(([ts]) => ts);
 
+      // Mirrors visualizationWidget.tsx — render the threshold mark areas /
+      // lines when the widget has at least one threshold value set.
+      const extraPlottables: Plottable[] = [];
+      const {thresholds} = data.widget;
+      if (
+        thresholds &&
+        (defined(thresholds.max_values.max1) || defined(thresholds.max_values.max2))
+      ) {
+        extraPlottables.push(
+          new Thresholds({
+            thresholds,
+            dataType: flatTimeSeries[0]?.meta?.valueType,
+          })
+        );
+      }
+
       return buildTimeseriesChartOption({
         theme,
         timeSeries: flatTimeSeries,
+        extraPlottables,
         createPlottable: (ts, {color}) => {
           const widgetQueryIndex = queryIndexByTimeSeries.get(ts) ?? 0;
           const widgetQuery =
